# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для университета",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/classes": {
            "get": {
                "tags": [
                    "Классы"
                ],
                "summary": "Получение списка классов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 3,
                                        "name": "02-СВ"
                                    },
                                    {
                                        "id": 4,
                                        "name": "05-ДВ"
                                    },
                                    {
                                        "id": 5,
                                        "name": "01-КГ"
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/api/classes/{id}": {
            "get": {
                "tags": [
                    "Классы"
                ],
                "summary": "Получение информации о конкретном классе (название, студенты класса)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id класса",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "3"
                                        },
                                        "name": {
                                            "type": "string",
                                            "maxLength": 255,
                                            "example": "02-СВ"
                                        },
                                        "students": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "Андрей Козлов"
                                                },
                                                {
                                                    "id": 9,
                                                    "name": "Виталий Дьяков"
                                                },
                                                {
                                                    "id": 11,
                                                    "name": "Андрей Афанасьев"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/students": {
            "get": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Получение списка студентов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Андрей Козлов",
                                        "email": "kozlov@mail.ru"
                                    },
                                    {
                                        "id": 2,
                                        "name": "Виктор Кулешов",
                                        "email": "kuleshov@mail.ru"
                                    },
                                    {
                                        "id": 4,
                                        "name": "Вячеслав Иванов",
                                        "email": "ivanov@mail.ru"
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Создание студента",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "name",
                                            "email",
                                            "student_class_id"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "pattern": "^[^0-9]*$",
                                                "example": "Иван Иванов"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "ivanov@mail.ru"
                                            },
                                            "student_class_id": {
                                                "type": "integer",
                                                "example": 3
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Студент успешно создан!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/students/{id}": {
            "get": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Получение информации о конкретном студенте (имя, email, класс, прослушанные лекции)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id студента",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "maxLength": 255,
                                            "pattern": "^[^0-9]*$",
                                            "example": "Иван Иванов"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "ivanov@mail.ru"
                                        },
                                        "student_class": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "02-СВ"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "lectures": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "topic": {
                                                        "type": "string",
                                                        "example": "Введение в ООП"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "В этой лекции вы познакомитесь с объектно ориентированным программированием на PHP. Вы узнаете о классах, интерфейсах, инкапсуляции и магических методах. В итоге поймете, для чего нужны классы как абстракция данных. Знание основ объектно ориентированного программирования пригодится, чтобы понимать плюсы и минусы кода с классами и объектами. Также это поможет использовать объекты для реализации программной логики. Этот курс подойдет тем, кто уже хорошо знаком с языком PHP. Чтобы учиться было проще, стоит заранее изучить курсы: «Основы PHP», «PHP: Массивы», «PHP: Ассоциативные массивы» и «PHP: Функции»."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Обновить информацию о студенте",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id студента",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "name",
                                            "email",
                                            "student_class_id"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "pattern": "^[^0-9]*$",
                                                "example": "Иван Ивановский"
                                            },
                                            "student_class_id": {
                                                "type": "integer",
                                                "example": 3
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Студент успешно обновлён!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Удалить студента",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id студента",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Студент успешно удалён!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}